backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2
django-cors-headers==4.7.0
django-oauth-toolkit==3.0.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
unidecode==1.4.0
urllib3==2.4.0
